Now delete_user should be working too
Rev: server/config_interface/standard/user_delete.pike:1.3
Rev: server/config_interface/standard/users.html:1.17

+++ server/config_interface/standard/user_delete.pike
@@ -1,14 +1,19 @@
+#include <roxen.h>
+//<locale-token project="config_interface"> LOCALE </locale-token>
+#define LOCALE(X,Y)  _STR_LOCALE("config_interface",X,Y)
+
 mixed parse( RequestID id )
 {
-  string res="";
+  string res="<br />";
   mapping v = id->variables;
   if(! id->misc->config_user->auth( "Edit Users" ) )
-    return "&var.no_perm_txt;";
+    return LOCALE("dy", "Permission denied");
 
   while( id->misc->orig ) id = id->misc->orig;
 
-  if( v->delete_user )
+  if( v->delete_user && v->delete_user!="")
   {
+    werror("\nGot %O\n",v->delete_user);
     id->misc->delete_old_config_user( v->delete_user );
     return "";
   }
@@ -16,16 +21,19 @@ mixed parse( RequestID id )
   {
     object u = id->misc->get_config_user( uid );
     if( u == id->misc->config_user )
-       res += ("<gbutton font='&usr.gbutton-font;' "
-	       "dim='1' width='300' preparse='1'> &var.delete_txt; "
-               + u->real_name+" ("+uid+")</gbutton><br />");
+      res += ("<gbutton font='&usr.gbutton-font;' "
+	      "dim='1' width='300' preparse='1'> " +
+	      LOCALE("dE", "Delete") +" "+ u->real_name+" ("+uid+") "
+	      "</gbutton><br />");
     else
-       res += ("<dbutton gbutton-width='300' "
-	       "gbutton-font='&usr.gbutton-font;' gbutton-preparse='1' "
-               "gbutton_title=' &var.delete_txt; "
-               + u->real_name + " ("+uid+")'/>"
-               " <insert file='user_delete.pike?delete_user="+
-	       Roxen.html_encode_string(uid)+"'/></dbutton><br />");
+      res += ("<dbutton gbutton-width='300' "
+	      "gbutton-font='&usr.gbutton-font;' "
+	      "gbutton-preparse='1' gbutton_title=' " +
+	      LOCALE("dE", "Delete") +" "+ u->real_name+" ("+uid+")' "
+	      "page='delete_user'>"
+	      "<insert file='user_delete.pike?delete_user="
+	      + Roxen.html_encode_string(uid) + 
+	      "'></dbutton><br />");
   }
   return res;
 }
